// zadanie2/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Rodzaj = {
  ZAKWATEROWANIE: 'ZAKWATEROWANIE',
  TRANSPORT: 'TRANSPORT',
  WYZYWIENIE: 'WYZYWIENIE'
} as const

export type Rodzaj = (typeof Rodzaj)[keyof typeof Rodzaj]


export const Dieta = {
  WEGETARIANSKA: 'WEGETARIANSKA',
  WEGANSKA: 'WEGANSKA',
  BEZGLUTENOWA: 'BEZGLUTENOWA',
  BEZLAKTOZOWA: 'BEZLAKTOZOWA'
} as const

export type Dieta = (typeof Dieta)[keyof typeof Dieta]


export const Miasto = {
  WARSZAWA: 'WARSZAWA',
  SEOUL: 'SEOUL',
  PARYZ: 'PARYZ',
  TOKIO: 'TOKIO',
  SZANGHAJ: 'SZANGHAJ',
  SINGAPUR: 'SINGAPUR'
} as const

export type Miasto = (typeof Miasto)[keyof typeof Miasto]


export const Wyzywienie = {
  ALL_INCLUSIVE: 'ALL_INCLUSIVE',
  POSILKI_3: 'POSILKI_3',
  POSILKI_2: 'POSILKI_2',
  SNIADANIA: 'SNIADANIA',
  BEZ_WYZYWIENIA: 'BEZ_WYZYWIENIA'
} as const

export type Wyzywienie = (typeof Wyzywienie)[keyof typeof Wyzywienie]